PrefectureComponentに都道府県のソート、更新差分の取得関数を移動
