+++ src/components/organisms/PrefectureComponent.vue
@@ -3,21 +3,47 @@ import '@/assets/css/prefectures/prefecture.css'
 import endpoint from '@/assets/ts/endpoint'
 import { Prefecture } from '@/assets/ts/interfaces/interfaces'
 import axios from '@/assets/plugins/setApiKey'
-import { ref, onMounted, watchEffect } from 'vue'
+import { ref, onMounted } from 'vue'
 
-interface Emits {
-    (emit: 'onSelectPrefecture', pref: Prefecture[]): void
-}
-
-const emit = defineEmits<Emits>()
+// interface Emits {
+//     (emit: 'onSelectPrefecture', pref: Prefecture[]): void
+// }
+// const emit = defineEmits<Emits>()
 
 // 都道府県一覧
 const prefectures = ref<Prefecture[]>([])
 // 選択された都道府県一覧
 const selectedPrefectures = ref<Prefecture[]>([])
+// 都道府県の人口情報
+const prefPopulation = ref<any>([])
 
-// 都道府県の選択状態を監視し、変更がなされた場合親コンポーネントに送る
-watchEffect(() => emit('onSelectPrefecture', selectedPrefectures.value))
+// 都道府県の更新差分を取得
+const getPrefDiffs = (
+    prevPrefs: Prefecture[],
+    newPrefs: Prefecture[]
+): Prefecture[] => {
+    return newPrefs.filter((pref) => prevPrefs.indexOf(pref) === -1)
+}
+
+// 都道府県をコード順にソートしたものを取得
+const sortPrefCodeAsc = (prefs: Prefecture[]): Prefecture[] => {
+    const sortedPrefs = prefs.sort((prev, next) => {
+        return prev.prefCode > next.prefCode ? 1 : -1
+    })
+    return sortedPrefs
+}
+
+// 都道府県の選択状態を監視し、変更がなされた場合一覧をソートし、更新差分を取得
+const onSelectPrefecture = (pref: Prefecture) => {
+    const prefIndex = selectedPrefectures.value.indexOf(pref)
+    if (prefIndex !== -1) {
+        selectedPrefectures.value.splice(prefIndex, 1)
+    } else {
+        selectedPrefectures.value.push(pref)
+        selectedPrefectures.value = sortPrefCodeAsc(selectedPrefectures.value)
+    }
+    console.log(selectedPrefectures.value)
+}
 
 // viewでlabelとinputの接続のためのidを作成する
 const setPrefId = (prefs: Prefecture[]): Prefecture[] => {
@@ -69,6 +95,7 @@ onMounted(async () => {
                     v-model="selectedPrefectures"
                     :value="prefecture"
                     :id="prefecture.prefId"
+                    @change="onSelectPrefecture(prefecture)"
                 />
                 <label :for="prefecture.prefId">
                     {{ prefecture.prefName }}
